refactor(console-ui): phase 5d — verification module + lazy cert-verify (F4/F7/F5)
Split the 653-line VerificationPanel into a components/verification/ module
(thin VerificationPanel orchestrator + useDeviceVerification hook + NormalMode /
TechnicalMode bodies + StatusLine/VerifyStepLine + types) and lazy-load the
heavy X.509 verifier:

- pkijs + asn1js (~76 KB gz) are no longer statically imported. The verifier is
  dynamically imported inside the verify action (useDeviceVerification, and the
  same minimal change in providers/dashboard/AttestationPanel); the types stay
  as erased type-only imports (perf F4).
- VerificationPanel + E2ELockIndicator now use the shared maskSerial (F5 — kills
  two of the four forked copies; identical output for real <=12-char serials).

Measured: chat route 360 -> 275 KB gz, providers route 323 -> 238 KB gz (both
-85 KB). Invariant held: build green, eslint 0 errors / 112 warnings, vitest 352
pass / 4 pre-existing fail. (stats still imports cert-verify — fixed in the
stats phase.)
