+++ console-ui/src/components/verification/types.ts
@@ -1,0 +1,9 @@
+// Provider attestation detail surfaced in the technical verification view.
+export interface ProviderDetail {
+  systemVolumeHash?: string;
+  sePublicKey?: string;
+  mdaSerial?: string;
+  mdaOsVersion?: string;
+  mdaSepVersion?: string;
+  mdaCertCount?: number;
+}
